LIBFCREPO-422. Create exceptions module.
Moved all exception classes into a class.exceptions module.

https://issues.umd.edu/browse/LIBFCREPO-422
