+++ classes/pcdm.py
@@ -6,6 +6,7 @@ import logging
 import threading
 from rdflib import Graph, Literal, URIRef
 from classes import ldp, ore
+from classes.exceptions import RESTAPIException
 from namespaces import dcterms, iana, pcdm, rdf
 
 # alias the RDFlib Namespace
@@ -225,12 +226,6 @@ class TransactionKeepAlive(threading.Thread):
 #============================================================================
 # PCDM RESOURCE (COMMON METHODS FOR ALL OBJECTS)
 #============================================================================
-
-class RESTAPIException(Exception):
-    def __init__(self, response):
-        self.response = response
-    def __str__(self):
-        return '{0} {1}'.format(self.response.status_code, self.response.reason)
 
 class Resource(ldp.Resource):
     def components(self):
